perf: inline sanitizeHtml into personalization, eliminate the module request
sanitizeHtml.js had a single consumer (personalization.js) and was always
loaded as a static import — meaning every page with personalization paid
for an extra serial module fetch. Moving the logic directly into
personalization.js eliminates that request entirely.

- Inline stringToHTML, isPossiblyDangerous, sanitizeHtmlNode, and
  sanitizeHtmlBody into personalization.js as private functions
  (sanitizeHtmlBody exported for testability)
- Delete libs/utils/sanitizeHtml.js
- Remove the now-unnecessary modulepreload hint for sanitizeHtml.js
- Update sanitizeHtml.test.js to import from personalization.js and
  drop the sanitizeHtml (firstChild) describe block — that export is gone
